+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- [docker]: See the guidelines on https://github.com/Luligu/matterbridge?tab=readme-ov-file#Run-the-Docker-container-and-start-it-with-full-access-to-the-host-network-and-with-volume-for-the-plugin-and-data-directory.
+- [docker]: See the guidelines on https://github.com/Luligu/matterbridge?tab=readme-ov-file#Run-the-Docker-container-and-start-it.
 - [docker]: Added support for docker. The Matterbridge image is published on the docker hub.
 
 ### Fixed